+++ crates/polars-plan/src/dsl/function_expr/mod.rs
@@ -968,31 +968,6 @@ impl From<FunctionExpr> for SpecialEq<Arc<dyn SeriesUdf>> {
     }
 }
 
-impl From<BinaryFunction> for SpecialEq<Arc<dyn SeriesUdf>> {
-    fn from(func: BinaryFunction) -> Self {
-        use BinaryFunction::*;
-        match func {
-            Contains => {
-                map_as_slice!(binary::contains)
-            },
-            EndsWith => {
-                map_as_slice!(binary::ends_with)
-            },
-            StartsWith => {
-                map_as_slice!(binary::starts_with)
-            },
-            #[cfg(feature = "binary_encoding")]
-            HexDecode(strict) => map!(binary::hex_decode, strict),
-            #[cfg(feature = "binary_encoding")]
-            HexEncode => map!(binary::hex_encode),
-            #[cfg(feature = "binary_encoding")]
-            Base64Decode(strict) => map!(binary::base64_decode, strict),
-            #[cfg(feature = "binary_encoding")]
-            Base64Encode => map!(binary::base64_encode),
-        }
-    }
-}
-
 #[cfg(feature = "temporal")]
 impl From<TemporalFunction> for SpecialEq<Arc<dyn SeriesUdf>> {
     fn from(func: TemporalFunction) -> Self {
